MB-59067: Move to 'supported-newer-1.20' golang branch
This is for use (hopefully exclusively) in couchbase-server 7.1.6 /
7.2.3 builds, which are quick turnaround releases with a few targeted
fixes. It just so happened that SUPPORTED_NEWER was bumped to 1.21
between the previous releases and those, which would result in a number
of components building with Go 1.21 which were on Go 1.20 in the
previous release. That's too significant a change for a quick patch
release. So this branch exists which has the newest versions, but still
says SUPPORTED_NEWER is 1.20.

Change-Id: Id0f3ae0b7dd900111feaeccf83ab8e2bf16a99fe
Reviewed-on: https://review.couchbase.org/c/manifest/+/198989
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>