MB-60538: Upgrade libcouchbase to include CCBC-1627 [1/2]
libcouchbase V3.3.11 generates invalid encoded packets
for unlock, get_replica and exists commands if the command
tried to use impersonate api added in CCBC-1216 as it
did not update the bodylength of the frame with the
length of the frame extras.

Pin to exact sha for trinity as the branch is restricted,
and the packages the current tagged versions of libcouchbase
generates the bug causing the failure. In order to get
server builds which don't use invalid encodings we need a
newer version of libcoucbase, and the SDK team don't want
to create a tag.

Change-Id: I81ef73b1d52852dbbc7fcffd7215a662a2852609
Reviewed-on: https://review.couchbase.org/c/manifest/+/204436
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>